Worked with Atheer and Lucas to begin setting up location table and functionality

+++ LocationFinder/src/main/java/LocationFinder/services/LocationService.java
@@ -14,7 +14,7 @@ public class LocationService {
     @Autowired
     private LocationRepository locRepository;
 
-    public List<Location> getByTemplate(Integer loc_id, String loc_name, String loc_area, Float loc_cost) {
+    public List<Location> getByTemplate(Integer loc_id, String loc_name, String loc_area, Double loc_cost) {
         return locRepository.findByTemplate(loc_id, loc_name, loc_area, loc_cost);
     }
 }
